Changed stuff in /network

+++ src/network/ip.h
@@ -4,7 +4,7 @@
 
 #include "../drivers/ethernet/enc28j60_instructions.h"
 
-void construct_ipv4_header(uint16_t data_length, uint8_t ttl, uint8_t protocol, int source_ip, int dest_ip);
+void construct_ipv4_header(uint32_t data_length, uint32_t ttl, uint32_t protocol, uint32_t source_ip, uint32_t dest_ip);
 uint16_t calculate_checksum(int * header);
 
 #include "ip.c"
